Add gitignore template and requirements, featurecollector
Feaurecollector scans through wav files in subdirectories of a "data" folder, then uses pyaudioanalysis to extract features from each wav file. The features are added to a pandas dataframe, and then it outputs a csv with the features for every file in the folder to the output folder. Next step will be to add plotting and such.